move entityExistsException in another package and register all mapper
