+++ back/src/main/java/org/superdev/coddy/JerseyConfig.java
@@ -3,8 +3,9 @@ package org.superdev.coddy;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.springframework.stereotype.Component;
 import org.superdev.coddy.application.api.Hello;
+import org.superdev.coddy.application.exception.mapper.EntityNotFoundMapper;
 import org.superdev.coddy.user.api.User;
-import org.superdev.coddy.user.exception.mapper.EntityExistsMapper;
+import org.superdev.coddy.application.exception.mapper.EntityExistsMapper;
 
 @Component
 public class JerseyConfig extends ResourceConfig {
@@ -16,5 +17,6 @@ public class JerseyConfig extends ResourceConfig {
         //register mapper
         register(SecurityManager.class);
         register(EntityExistsMapper.class);
+        register(EntityNotFoundMapper.class);
     }
 }
